Fix for issue 736.
SECTION: Modified Files
----------------------------
M jsf-ri/src/com/sun/faces/taglib/jsf_core/ValidateLongRangeTag.java
 - use Long.valueOf() for literal expressions instead of Integer.valueOf().

M jsf-ri/systest/web/validator02.jsp
 - validate the issue/fix

git-svn-id: https://svn.java.net/svn/mojarra~svn/trunk@4787 761efcf2-d34d-6b61-9145-99dcacc3edf1

+++ jsf-ri/src/com/sun/faces/taglib/jsf_core/ValidateLongRangeTag.java
@@ -149,7 +149,7 @@ public class ValidateLongRangeTag extends MaxMinValidatorTag {
                                   context)).longValue();
             } else {
                 minimum =
-                Integer.valueOf(minimumExpression.getExpressionString()).
+                Long.valueOf(minimumExpression.getExpressionString()).
                     longValue();
             }
         }
@@ -160,7 +160,7 @@ public class ValidateLongRangeTag extends MaxMinValidatorTag {
                                   context)).longValue();
             } else {
                 maximum =
-                Integer.valueOf(maximumExpression.getExpressionString()).
+                Long.valueOf(maximumExpression.getExpressionString()).
                     longValue();
             }
         }
